basic file input

+++ src/gsp14/mips_cpu.h
@@ -16,7 +16,7 @@ void cpu_decode_instr(const uint32_t &bytes,uint32_t &op,uint32_t &s,uint32_t &t
 
 mips_error cpu_execute_r(uint32_t src1, uint32_t src2, uint32_t dest, uint32_t shift_amt, uint32_t fn_code, mips_cpu_h state);
 mips_error cpu_execute_rt(const uint32_t &src,const uint32_t &fn_code, const uint16_t &i, mips_cpu_h state);
-mips_error cpu_execute_i(const uint32_t &s,const uint32_t &t, const uint32_t &i, const uint32_t &opcode, mips_cpu_h state);
+mips_error cpu_execute_i(const uint32_t &s,const uint32_t &t, const uint16_t &i, const uint32_t &opcode, mips_cpu_h state);
 mips_error cpu_execute_j(const uint32_t &j,const uint32_t opcode, mips_cpu_h state);
 
 void less_than(uint32_t src1, uint32_t src2, uint32_t dest, uint32_t fn_code,mips_cpu_h state);
